pass openssl values in nativeland

+++ packages/bun-usockets/src/libusockets.h
@@ -239,8 +239,8 @@ struct us_bun_socket_context_options_t {
     int request_cert;
     unsigned int client_renegotiation_limit;
     unsigned int client_renegotiation_window;
-    double min_tls_version;
-    double max_tls_version;
+    unsigned int min_tls_version;
+    unsigned int max_tls_version;
 };
 
 /* Return 15-bit timestamp for this context */
